Decimal money (Again...)

+++ ProjectsInfo/Models/Employes/Employee.cs
@@ -5,10 +5,10 @@ namespace ProjectsInfo.Models.employes
     public class Employee
     {
         // Обшие затраты - константа, прибавляющаяся к стоимости часа
-        private const double GenralExpences = 100;
+        private const decimal GenralExpences = 100;
 
         // Повышающий коэффициент
-        private const double Multiplier = 2.5;
+        private const decimal Multiplier = 2.5m;
 
         // Primary key
         public int ID { get; set; }
@@ -18,9 +18,9 @@ namespace ProjectsInfo.Models.employes
 
         [Display(Name = "Зарплата")]
         [DataType(DataType.Currency)]
-        public double Salary { get; set; }
+        public decimal Salary { get; set; }
 
         [Display(Name = "Стоимость часа работы")]
-        public double HourPrice => (Salary / 160) * Multiplier + GenralExpences;
+        public decimal HourPrice => (Salary / 160) * Multiplier + GenralExpences;
     }
 }
